refactor: move second level commands to their own packages
