+++ internal/registry/auth/basic/doc.go
@@ -1,4 +1,4 @@
-// Copyright 2023 The Falco Authors
+// Copyright 2022 The Falco Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,5 +12,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package auth defines the logic to authenticate against an OCI registry.
-package auth
+// Package basic defines the logic to authenticate against an OCI registry.
+package basic
